requirements.txt
flask==3.1.0
python-dotenv==1.0.1
requests==2.32.3
APScheduler==3.10.4
anthropic==0.42.0
gunicorn==23.0.0
